Splitting up the view respositories
