+++ src/Resources/config/services/queries.xml
@@ -1,50 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
-    <services>
-        <defaults public="true" />
+<container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+           xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
 
-        <service id="sylius.shop_api_plugin.view_repository.cart_view_repository" class="Sylius\ShopApiPlugin\ViewRepository\CartViewRepository">
-            <argument type="service" id="sylius.repository.order" />
-            <argument type="service" id="sylius.shop_api_plugin.factory.cart_view_factory" />
-        </service>
-
-        <service id="sylius.shop_api_plugin.view_repository.placed_order_view_repository" class="Sylius\ShopApiPlugin\ViewRepository\PlacedOrderViewRepository">
-            <argument type="service" id="sylius.repository.order" />
-            <argument type="service" id="sylius.repository.customer" />
-            <argument type="service" id="sylius.shop_api_plugin.factory.placed_order_view_factory" />
-        </service>
-
-        <service id="sylius.shop_api_plugin.view_repository.product_details_view_repository" class="Sylius\ShopApiPlugin\ViewRepository\ProductDetailsViewRepository">
-            <argument type="service" id="sylius.repository.channel" />
-            <argument type="service" id="sylius.repository.product" />
-            <argument type="service" id="sylius.shop_api_plugin.factory.detailed_product_view_factory" />
-            <argument type="service" id="sylius.shop_api_plugin.provider.supported_locale_provider"/>
-        </service>
-
-        <service id="sylius.shop_api_plugin.view_repository.product_catalog_view_repository" class="Sylius\ShopApiPlugin\ViewRepository\ProductCatalogViewRepository">
-            <argument type="service" id="sylius.repository.channel" />
-            <argument type="service" id="sylius.repository.product" />
-            <argument type="service" id="sylius.repository.taxon" />
-            <argument type="service" id="sylius.shop_api_plugin.factory.list_product_view_factory" />
-            <argument type="service" id="sylius.shop_api_plugin.factory.page_view_factory" />
-            <argument type="service" id="sylius.shop_api_plugin.provider.supported_locale_provider"/>
-        </service>
-
-        <service id="sylius.shop_api_plugin.view_repository.product_reviews_view_repository" class="Sylius\ShopApiPlugin\ViewRepository\ProductReviewsViewRepository">
-            <argument type="service" id="sylius.repository.channel" />
-            <argument type="service" id="sylius.repository.product_review" />
-            <argument type="service" id="sylius.repository.product" />
-            <argument type="service" id="sylius.shop_api_plugin.factory.product_review_view_factory" />
-            <argument type="service" id="sylius.shop_api_plugin.factory.page_view_factory" />
-            <argument type="service" id="sylius.shop_api_plugin.provider.supported_locale_provider"/>
-        </service>
-
-        <service id="sylius.shop_api_plugin.view_repository.product_latest_view_repository" class="Sylius\ShopApiPlugin\ViewRepository\ProductLatestViewRepository">
-            <argument type="service" id="sylius.repository.channel" />
-            <argument type="service" id="sylius.repository.product" />
-            <argument type="service" id="sylius.shop_api_plugin.factory.list_product_view_factory" />
-            <argument type="service" id="sylius.shop_api_plugin.provider.supported_locale_provider"/>
-        </service>
-    </services>
+    <imports>
+        <import resource="repositories/**/*"/>
+    </imports>
 </container>
