+++ src/Resources/config/services/repositories/order.xml
@@ -1,0 +1,12 @@
+<container xmlns="http://symfony.com/schema/dic/services">
+    <services>
+        <defaults public="true" />
+
+        <service id="sylius.shop_api_plugin.view_repository.placed_order_view_repository"
+                 class="Sylius\ShopApiPlugin\ViewRepository\Order\PlacedOrderViewRepository">
+            <argument type="service" id="sylius.repository.order" />
+            <argument type="service" id="sylius.repository.customer" />
+            <argument type="service" id="sylius.shop_api_plugin.factory.placed_order_view_factory" />
+        </service>
+    </services>
+</container>
