+++ src/ViewRepository/Order/PlacedOrderViewRepositoryInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Sylius\ShopApiPlugin\ViewRepository;
+namespace Sylius\ShopApiPlugin\ViewRepository\Order;
 
 use Sylius\ShopApiPlugin\View\PlacedOrderView;
